feature_BGDIINF_SB-2328: renamed clientdata_dynamodb.py to s3_client.py
As no dynamodb part is in that module anymore.
